refactor(imports): adjust import paths to match updated naming
