+++ src/components/task-board/task-sheet/delete-task.tsx
@@ -9,7 +9,7 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog";
 import { useTaskActions } from "@/hooks/use-task-actions";
-import { useTaskSheet } from "@/store/use-task-sheet-store";
+import { useTaskSheet } from "@/store/useTaskSheet";
 import { DialogClose } from "@radix-ui/react-dialog";
 import { Trash } from "lucide-react";
 import toast from "react-hot-toast";
