+++ src/components/task-board/task-sheet/task-sheet-content.tsx
@@ -2,7 +2,6 @@ import DeleteTask from "./delete-task";
 import { Button } from "@/components/ui/button";
 import { ArrowRight } from "lucide-react";
 import MarkCompleteButton from "./mark-complete-button";
-import { useTaskSheet } from "@/store/use-task-sheet-store";
 import React, { useCallback } from "react";
 import TaskSheetName from "./task-sheet-name";
 import TaskSheetStatus from "./task-sheet-status";
@@ -10,6 +9,7 @@ import TaskSheetDate from "./task-sheet-date";
 import TaskSheetAssignee from "./task-sheet-assignee";
 import TaskSheetPriority from "./task-sheet-priority";
 import TaskSheetDescription from "./task-sheet-description";
+import { useTaskSheet } from "@/store/useTaskSheet";
 
 const TaskSheetContent = () => {
   const { task, setIsOpen, resetTask } = useTaskSheet();
